게임파일 정리, 2048게임 기반 마련
