+++ pythonBasic/Games/game2048.py
@@ -1,0 +1,2 @@
+def run2048():
+    print("2048 시작!")
